move files into src components and change names
