+++ App.js
@@ -1,7 +1,7 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import AllTabs from './Components/AllTabs';
+import AllTabs from './src/components/AllTabs'
 
 
 export default function App() {
